requirements.txt
alabaster==0.7.12
atomicwrites==1.3.0
attrs==19.1.0
Babel==2.6.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
colorama==0.4.1
convertdate==2.1.3
coverage==4.5.3
cycler==0.10.0
Cython==0.29.6
docutils==0.14
ephem==3.7.6.0
fbprophet==0.4.post2
holidays==0.9.10
hypothesis==4.14.2
idna==2.8
imagesize==1.1.0
Jinja2==2.10
JPype1==0.6.3
kiwisolver==1.0.1
lunardate==0.2.0
MarkupSafe==1.1.1
matplotlib==3.0.3
more-itertools==7.0.0
numpy==1.16.2
packaging==19.0
pandas==0.24.2
pluggy==0.9.0
py==1.8.0
pyfiglet==0.8.post1
Pygments==2.3.1
pyparsing==2.3.1
pystan==2.18.1.0
pytest==4.4.0
pytest-cov==2.6.1
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
responses==0.10.6
scipy==1.2.1
setuptools-git==1.2
six==1.12.0
snowballstemmer==1.2.1
Sphinx==2.0.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.1
sphinxcontrib-websupport==1.1.0
tqdm==4.31.1
urllib3==1.24.1
